% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataArray.R
\name{CdataArray}
\alias{CdataArray}
\title{Compute data array}
\usage{
CdataArray(distanceMatrices, ineqmembers)
}
\arguments{
\item{distanceMatrices}{A list of arrays of distance values, one
for each market. See the appropriate section for their definition.}

\item{ineqmembers}{A list of objects describing the indices participating in
the inequalities occurring in the objective function. See the appropriate
section for their definition.}
}
\value{
An array of dimension \code{(noAttr, sumNumIneqs)}, where
\code{noAttr} is the number of distance attributes and \code{sumNumIneqs}
is the sum of the number of inequalities of each market over all markets.
\code{[paramIdx, ineqIdx]} indexes the \code{paramIdx}-th attribute for the
\code{ineqIdx}-th inequality.
}
\description{
Computes the data array, as defined in section 5.2 of [1].
}
\details{
Each column of the data array corresponds to a single inequality. To compute
its value, we gather the index triples (for market, upstream, and downstream)
appearing on either side of the inequality, and we generate the corresponding
term (a column vector of distance attribute values) from the distance
matrices. The difference between the sum of the LHS and RHS terms is equal to
the desired column.
}
\section{Old code}{

To port code using the previous version, replace \code{dataArray[[ineqIdx]]}
and \code{dataArray[[ineqIdx]][paramIdx]} by \code{dataArray[, ineqIdx]} and
\code{dataArray[paramIdx, ineqIdx]} respectively.
}

\section{References}{

\tabular{ll}{
[1] \tab David Santiago and Fox, Jeremy (2009). "A Toolkit for Matching
Maximum Score Estimation and Point and Set Identified Subsampling
Inference".
}
}

\section{Distance matrix structure}{

Let \code{mIdx} index a market. Each \code{distanceMatrix} is an array
(technically not a matrix) of dimension \code{(noAttr, noD[mIdx],
noU[mIdx])}. The element indexed by \code{[i, dIdx, uIdx]} gives the
\code{i}-th distance attribute value for the triple \code{(mIdx, uIdx,
dIdx)}.
}

\section{Inequality members structure}{

The inequality members structure is list with members:
\tabular{ll}{
\code{$fctUpIdxs} \tab A list of upstream index vectors for the factual case.\cr
\code{$fctDnIdxs} \tab A list of downstream index vectors for the factual case.\cr
\code{$cfcUpIdxs} \tab A list of upstream index vectors for the counterfactual case.\cr
\code{$cfcDnIdxs} \tab A list of downstream index vectors for the counterfactual case.\cr
\code{$numIneqs}  \tab The number of inequalities for this market. Also the
common length of the above lists.
}
Each list element of the first four members corresponds to a single
inequality. That element is a vector of indices, with one index for each
term in that inequality.
}

