% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{checkHeaderMatched}
\alias{checkHeaderMatched}
\title{Check if header is valid}
\usage{
checkHeaderMatched(header, fieldMode)
}
\arguments{
\item{header}{A vector of field names to be checked.}

\item{fieldMode}{A string denoting how to identify the fields of the file.
Options are:
\tabular{ll}{
\code{"position"} \tab (default) Identify each field based on its order
in the field list. \cr
\code{"name"} \tab Identify each field based on its name.
}
See the section "File structure" for more information.}
}
\description{
Checks if the given vector of field names satisfies the conditions stated in
\code{importMatched}, stopping execution if they are not met.
}
\keyword{internal}
