% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{checkIndices}
\alias{checkIndices}
\title{Check if indices are valid}
\usage{
checkIndices(idxs)
}
\arguments{
\item{idxs}{A vector of index values to be checked.}
}
\description{
Checks if the given vector of index values is non-empty and consecutive,
with values starting from \code{1}. Raises a warning otherwise.
}
\keyword{internal}
