% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalities.R
\name{Cineqmembers}
\alias{Cineqmembers}
\title{Compute inequality members for all markets}
\usage{
Cineqmembers(mate)
}
\arguments{
\item{mate}{See the return value of \code{importMatched}.}
}
\value{
A list of structures for each market. See the appropriate section for
its definition.
}
\description{
Computes the upstream and downstream indices for each inequality term for all
markets.
}
\section{Inequality members structure}{

The inequality members structure is list with members:
\tabular{ll}{
\code{$fctUpIdxs} \tab A list of upstream index vectors for the factual case.\cr
\code{$fctDnIdxs} \tab A list of downstream index vectors for the factual case.\cr
\code{$cfcUpIdxs} \tab A list of upstream index vectors for the counterfactual case.\cr
\code{$cfcDnIdxs} \tab A list of downstream index vectors for the counterfactual case.\cr
\code{$numIneqs}  \tab The number of inequalities for this market. Also the
common length of the above lists.
}
Each list element of the first four members corresponds to a single
inequality. That element is a vector of indices, with one index for each
term in that inequality.
}

