% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalities.R
\name{CineqmembersSingle}
\alias{CineqmembersSingle}
\title{Compute inequality members for a single market}
\usage{
CineqmembersSingle(marketMates)
}
\arguments{
\item{marketMates}{A list with members:
\tabular{ll}{
\code{$UpStream} \tab A vector of the unique indices of upstreams for this
market.\cr
\code{$DownMates} \tab A list with n elements, where n is the number of
upstreams. Element \code{$DownMates[[i]]} is a vector of the downstream
indices which the \code{i}-th upstream is matched to.
}}
}
\value{
See the section "Inequality members structure".
}
\description{
Computes the upstream and downstream indices for each inequality term for a
single market.
}
\details{
Let \eqn{n} be the number of upstreams in the market. Each transposition
\eqn{(i, j)} of the set \eqn{{1, 2, ..., n}} yields a single inequality, with
the factual terms in the LHS and the counterfactual terms in the RHS. The
factual terms correspond to all current matches of both upstream indices
(\eqn{i} and \eqn{j}), and the counterfactual terms correspond to the matches
where \eqn{i} and \eqn{j} switch partners.
}
\section{Inequality members structure}{

The inequality members structure is list with members:
\tabular{ll}{
\code{$fctUpIdxs} \tab A list of upstream index vectors for the factual case.\cr
\code{$fctDnIdxs} \tab A list of downstream index vectors for the factual case.\cr
\code{$cfcUpIdxs} \tab A list of upstream index vectors for the counterfactual case.\cr
\code{$cfcDnIdxs} \tab A list of downstream index vectors for the counterfactual case.\cr
\code{$numIneqs}  \tab The number of inequalities for this market. Also the
common length of the above lists.
}
Each list element of the first four members corresponds to a single
inequality. That element is a vector of indices, with one index for each
term in that inequality.
}

\keyword{internal}
