% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalities.R
\name{Cinequalities}
\alias{Cinequalities}
\title{Evaluate inequalities}
\usage{
Cinequalities(payoffFunction, ineqmembers)
}
\arguments{
\item{payoffFunction}{A function which takes three arguments
(\code{mIdx}, \code{uIdx}, \code{dIdx}) and returns the value of the payoff
function for that triple.}

\item{ineqmembers}{The output of \code{Cineqmembers}.}
}
\value{
A list of vectors, one for each market. Each vector element
corresponds to a single inequality.
}
\description{
Computes for each inequality the value
}
