% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{CmatchMatrices}
\alias{CmatchMatrices}
\title{Calculate optimal matchings for all markets}
\usage{
CmatchMatrices(payoffMatrices, quotasU, quotasD)
}
\arguments{
\item{payoffMatrices}{TODO See \code{evaluatePayoffMatrices}.}

\item{quotasU, quotasD}{TODO}
}
\value{
TODO
}
\description{
TODO docs
}
