% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R
\name{evaluatePayoffMatrices}
\alias{evaluatePayoffMatrices}
\title{TODO}
\usage{
evaluatePayoffMatrices(unevalPayoffMatrices, beta)
}
\arguments{
\item{unevalPayoffMatrices}{TODO}

\item{beta}{TODO}
}
\value{
TODO
}
\description{
TODO
}
