% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R
\name{evaluatePayoffMatrix}
\alias{evaluatePayoffMatrix}
\title{TODO}
\usage{
evaluatePayoffMatrix(unevalPayoffMatrix, beta)
}
\arguments{
\item{unevalPayoffMatrix}{TODO}

\item{beta}{TODO}
}
\value{
TODO
}
\description{
TODO
}
