% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{extractMate}
\alias{extractMate}
\title{Extract mate tables from imported table}
\usage{
extractMate(marketData)
}
\arguments{
\item{marketData}{The return value of \code{importCommon}.}
}
\value{
A list of mate tables.
}
\description{
Extract mate tables from imported table
}
\section{Mate table structure}{

Let \code{mIdx} index a market. Each mate table is a \code{data.table} object
with fields \code{UpStream} and \code{DownMates}. Each upstream in the market
corresponds to a row, whose \code{DownMates} field contains a vector of the
downstream indices which the upstream is matched to. The rows are ordered
consecutively based on the upstream index \code{UpStream}.
}

\keyword{internal}
