% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{extractQuotas}
\alias{extractQuotas}
\title{Extract quota vectors from imported table}
\usage{
extractQuotas(marketData)
}
\arguments{
\item{marketData}{The return value of \code{importUnmatchedMain}.}
}
\value{
A list of quota vectors, on for each market.
}
\description{
Extract quota vectors from imported table
}
\keyword{internal}
