% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalities.R
\name{getAllIneqTermsIdxs}
\alias{getAllIneqTermsIdxs}
\title{Get indices for all inequalities}
\usage{
getAllIneqTermsIdxs(ineqmembersSingle)
}
\arguments{
\item{ineqmembersSingle}{An element of the list constructed by
\code{Cineqmembers},corresponding to a single market.}
}
\value{
A list of arrays, one for each inequality. For the structure of those
arrays, see the return value of \code{getIneqTermsIdxs}.
}
\description{
Similar to \code{getIneqTermsIdxs}, but works on all inequalities. A
convenience function.
}
