% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximize.R
\name{getDefaultOptimParams}
\alias{getDefaultOptimParams}
\title{Get default values for the \code{optimParams} argument of \code{optimizeScoreFunction}}
\usage{
getDefaultOptimParams(method = NULL)
}
\arguments{
\item{method}{(optional) A string denoting the optimization method.
Currently, only \code{"DEoptim"} (the default value) is supported.}
}
\value{
A list of parameters.
}
\description{
Get default values for the \code{optimParams} argument of \code{optimizeScoreFunction}
}
