% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalities.R
\name{getIneqTermsIdxs}
\alias{getIneqTermsIdxs}
\title{Get indices for inequality}
\usage{
getIneqTermsIdxs(ineqmembersSingle, ineqIdx)
}
\arguments{
\item{ineqmembersSingle}{An element of the list constructed by
\code{Cineqmembers},corresponding to a single market.}

\item{ineqIdx}{The index of the inequality within that market.}
}
\value{
An array of dimension \code{(numTerms, 4)}, where \code{numTerms} is
the number of terms the given inequality has. The columns of the array have
names \code{fctUpIdxs}, \code{fctDnIdxs}, \code{cfcUpIdxs},
\code{cfcDnIdxs}. For information on those names, see the documentation of
\code{Cineqmembers}.
}
\description{
Returns, for a given inequality, the 4-tuple of indices defining each term in
that inequality. A convenience function.
}
