% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{loglikelihoodCommon}
\alias{loglikelihoodCommon}
\title{Common calculations for ROT log-likelihood}
\usage{
loglikelihoodCommon(y, x, beta, gamma)
}
\arguments{
\item{y}{TODO}

\item{x}{TODO}

\item{beta}{TODO}

\item{gamma}{TODO}
}
\value{
TODO
}
\description{
Common calculations for ROT log-likelihood
}
\keyword{internal}
