% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{loglikelihoodFixedBeta}
\alias{loglikelihoodFixedBeta}
\title{ROT log-likelihood with fixed beta}
\usage{
loglikelihoodFixedBeta(y, x, betaEst, par)
}
\arguments{
\item{y}{TODO}

\item{x}{TODO}

\item{par}{TODO}
}
\value{
TODO
}
\description{
ROT log-likelihood with fixed beta
}
\keyword{internal}
