% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{makeBootstrapExtraObjFun}
\alias{makeBootstrapExtraObjFun}
\title{Create objective function used in the \code{\link{cubeRootBootstrapCR}}
(with extra information)}
\usage{
makeBootstrapExtraObjFun(
  fullDataArray,
  sampleDataArray,
  betaEst,
  H,
  coefficient1 = 1,
  objSign = -1,
  useCorrectionFactor = TRUE
)
}
\arguments{
\item{fullDataArray}{The full data array.}

\item{sampleDataArray}{The sample data array.}

\item{betaEst}{The estimate of \eqn{\beta}. This should have been obtained
by calling \code{optimizeScoreFunction} on \code{fullDataArray}.}

\item{H}{The matrix \eqn{H}, as defined in the paper by Cattaneo et al..
Should have
dimensions \mjseqn{d \times d}, where \eqn{d} is the number of free
parameters.}

\item{coefficient1}{The first coefficient in the extended vector of
parameters. Should typically be \code{1} (the default) or \code{-1}.}

\item{objSign}{Use \code{-1} (the default) when passing the result to an
optimization routine that can only minimize, such as
\code{DEoptim::DEoptim}. Otherwise, use \code{1}.}

\item{useCorrectionFactor}{A boolean selecting whether to apply the
correction factor. If \code{TRUE} (the default value), then
\mjseqn{c = \frac{m}{n}}, otherwise it is \mjseqn{1}.}
}
\value{
A function taking a vector \code{b} and returning a list with members:
\tabular{ll}{
\verb{$val} \tab The value that the result of \code{\link{makeScoreObjFun}} would
return, if called with argument with argument \code{b}. \cr
\verb{$arg} \tab The argument \code{b}. \cr
\verb{$diffArg} \tab The value \mjseqn{\beta - \hat{\beta}}. \cr
\verb{$terms} \tab The values of the three terms of \mjseqn{B}, as a vector.
}
}
\description{
Similar to \code{\link{makeBootstrapObjFun}}, except it also returns extra
information on the terms of \eqn{B}.
}
\details{
See \code{\link{makeBootstrapObjFun}}.

\loadmathjax
}
\keyword{internal}
