% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R
\name{makeDistanceMatrices}
\alias{makeDistanceMatrices}
\title{Convert attribute matrices to distance matrices}
\usage{
makeDistanceMatrices(attributeMatricesUp, attributeMatricesDn)
}
\arguments{
\item{attributeMatricesUp, attributeMatricesDn}{See the return value of the
function \code{importUnmatched}.}
}
\value{
See the return value of the function \code{importMatched}.
}
\description{
Calculates the distance matrices for all markets, given data for unmatched
markets (the attribute matrices for upstreams and downstreams).
}
\details{
Let \code{mIdx}, \code{uIdx}, \code{dIdx}, and \code{k} index markets,
upstreams, downstreams, and attributes respectively. The distance matrix
values are defined by the equation:

\code{distanceMatrices[[mIdx]][k, dIdx, uIdx] ==
  attributeMatricesUp[[mIdx]][k, uIdx] * attributeMatricesDn[[mIdx]][k, dIdx]}.
}
