% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{makeHnumder}
\alias{makeHnumder}
\title{\loadmathjax
Create estimator for H matrix}
\usage{
makeHnumder(dataArray, betaEst, eps)
}
\arguments{
\item{dataArray}{The output of \code{dataArray}.}

\item{betaEst}{The vector \mjseqn{\hat{\beta}}, of length \mjseqn{d}, whose
value is the estimate of \mjseqn{\beta}, obtained from maximizing the
score function corresponding to \code{dataArray}.}

\item{eps}{The numerical derivative step size matrix
\mjseqn{\epsilon_{n,kl}}, of size \mjseqn{d \times d}.}
}
\value{
The matrix \mjseqn{H}.
}
\description{
Creates the estimator matrix \mjseqn{\tilde{H}_n} using the numerical
derivative method.
}
\keyword{internal}
