% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{makeHplugin}
\alias{makeHplugin}
\title{\loadmathjax}
\usage{
makeHplugin(dataArray, betaEst, h)
}
\arguments{
\item{dataArray}{The output of \code{dataArray}.}

\item{betaEst}{The vector \mjseqn{\hat{\beta}}, of length \mjseqn{d}, whose
value is the estimate of \mjseqn{\beta}, obtained from maximizing the
score function corresponding to \code{dataArray}.}

\item{h}{The bandwidth matrix
\mjseqn{h_{n,kl}}, of size \mjseqn{d \times d}.}
}
\value{
The matrix \mjseqn{H}.
}
\description{
Creates the estimator matrix \mjseqn{\tilde{H}_n} using the plug-in (kernel)
method.
}
\details{
\subsection{Plug-in kernel}{

Note that the plug-in estimator method requires a kernel function with
specific properties, as shown in the paper and its supplement. The only
kernel currently available is the function \mjseqn{K(u) = \phi(u)}, i.e. the
pdf of the standard normal distribution, with first derivative
\mjseqn{\dot{K}(u) = -u\phi(u)}.
}
}
