% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{makeObjFun}
\alias{makeObjFun}
\title{Create objective score function}
\usage{
makeObjFun(dataArray, coefficient1 = 1, objSign = -1)
}
\arguments{
\item{dataArray}{TODO See \code{CdataArray}.}

\item{coefficient1}{The first coefficient in the extended vector of
parameters. Should typically be \code{1} (the default) or \code{-1}.}

\item{objSign}{Use \code{-1} (the default) when passing the result to an
optimization routine that can only minimize, such as
\code{DEoptim::DEoptim}. Otherwise, use \code{1}.}
}
\value{
A function taking a vector of length \code{noAttr-1} (the parameters)
and returning an integer (the score).
}
\description{
Creates a score function from the given data array, in a form suitable for
passing to an optimization routine. This function computes the score (number
of inequalities satisfied) for a given parameter vector (\eqn{\beta}).
}
\details{
The function created is:
\eqn{g(\beta) = \mathrm{objSign} \sum_{k=1}^{N_{\mathrm{ineqs}} (a_k \beta')}}
where
\enumerate{
\item \eqn{\beta} is the vector of free parameters, of length \code{noAttr-1}.
\item \eqn{\beta' = (\mathrm{coefficient1}\, \beta)} is the extended
vector of parameters.
\item \eqn{a_k} is the \eqn{k}-th column of \code{dataArray}.
}
}
