% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximize.R
\name{maximizeDEoptim}
\alias{maximizeDEoptim}
\title{Helper function for Differential Evolution}
\usage{
maximizeDEoptim(objective, lower, upper, control = list())
}
\arguments{
\item{objective}{The objective function, produced by \code{makeScoreObjFun}.}

\item{lower, upper}{The lower and upper bounds in the box constraints.}

\item{control}{A list passed to \code{DEoptim::DEoptim}. See also
\code{DEoptim::DEoptim.control}.}
}
\value{
A list with elements:
\tabular{ll}{
\code{$optVal} \tab The optimal value of the objective function. \cr
\code{$optArg} \tab The argument for which the objective function attains
that value, as a vector.
}
}
\description{
An adapter function over \code{DEoptim::DEoptim}. When other methods are
implemented, they should follow this signature (for the three required
arguments).
}
\keyword{internal}
