% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{mergeOptions}
\alias{mergeOptions}
\title{Merge passed options with default options}
\usage{
mergeOptions(options, defaultOptions)
}
\arguments{
\item{options}{A list of name-value pairs.}

\item{defaultOptions}{A list of name-value pairs.}
}
\value{
A list of name-value pairs obtained by adding to \code{options} any pairs
only occuring in \code{defaultOptions}.
}
\description{
Merges the passed options list with the default options list by adding to
the former any name-value pairs which only occur in the latter. In case of
name conflicts, the former pair is kept.
}
\examples{
\donttest{mergeOptions(list(x = 1, y = 2), list(y = 3, z = 4))}
}
\keyword{internal}
