% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{newBootstrapCR}
\alias{newBootstrapCR}
\title{Calculate confidence region}
\usage{
newBootstrapCR(
  dataArray,
  groupIDs,
  pointEstimate,
  ssSize,
  numSubsamples,
  confidenceLevel,
  optimizeScoreArgs,
  options = NULL
)
}
\arguments{
\item{dataArray}{The output of \code{dataArray}.}

\item{groupIDs}{The output of \code{makeGroupIDs}.}

\item{pointEstimate}{The optimal parameters, as calculated by
\code{optimizeScoreFunction}.}

\item{ssSize}{The size of the market subset used in subsampling. Must not be
greater than the number of markets.}

\item{numSubsamples}{The number of subsamples.}

\item{confidenceLevel}{The confidence level of the region. Must be a number
in the range \eqn{(0, 1)}.}

\item{optimizeScoreArgs}{A list with the keyword arguments to be used when
\code{optimizeScoreFunction} is called. All non-optional arguments should
be present, except \code{dataArray}.}

\item{options}{A list of options:
\tabular{ll}{
\code{progressUpdate}  \tab How often to print progress. Defaults to
\code{0} (never). \cr
\code{asymptotics}     \tab Type of asymptotics to use. Supported values
are "nests" (default) or "coalitions".
}}
}
\value{
TODO
}
\description{
Generates a confidence region estimate using Cattaneo's bootstrap method.
}
\details{
See:
https://cattaneo.princeton.edu/papers/Cattaneo-Jansson-Nagasawa_2020_ECMA.pdf
https://cattaneo.princeton.edu/papers/Cattaneo-Jansson-Nagasawa_2020_ECMA--Supplement.pdf
https://github.com/mdcattaneo/replication-CJN_2020_ECMA
}
