% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximize.R
\name{optimizeScoreFunction}
\alias{optimizeScoreFunction}
\title{Calculate maximum score}
\usage{
optimizeScoreFunction(
  dataArray,
  bounds,
  coefficient1 = NULL,
  method = NULL,
  optimParams = NULL,
  getIneqSat = FALSE,
  permuteInvariant = FALSE
)
}
\arguments{
\item{dataArray}{The output of \code{CdataArray}.}

\item{bounds}{A list with elements \code{$lower}, \code{$upper} which are
vectors defining lower and upper bounds for each variable in the objective
function.}

\item{coefficient1}{(optional) The first coefficient of the extended
parameter vector. See \code{makeObjFun} for more details.}

\item{method}{(optional) A string denoting the optimization method.
Currently, only \code{"DEoptim"} (the default value) is supported.}

\item{optimParams}{(optional) A list of parameters to be passed to the
optimization routine. Defaults to the empty list (TODO see \code{do.call}).}

\item{getIneqSat}{(optional) A boolean indicating whether to include the
\code{$ineqSat} member in the result. Defaults to \code{FALSE}.}

\item{permuteInvariant}{(optional) TODO. Defaults to \code{FALSE}.}
}
\value{
A list with members:
\tabular{ll}{
\code{$optVal}  \tab The optimal value of the objective function.\cr
\code{$optArg}  \tab The argument vector which achieves that value.\cr
\code{$ineqSat} \tab A vector of \code{0}s and \code{1}s. Each element
corresponds to a single inequality, and is \code{1} if that inequality is
satisfied for the optimal parameters, and \code{0} otherwise. Only
present if \code{getIneqSat} is \code{TRUE}.
}
}
\description{
Optimizes the maximum score function defined by the data array. The objective
function is created using \code{makeObjFun}.
}
\section{Optimization methods}{

TODO
See \code{DEoptim::DEoptim.control}.
}

