% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{plotCR}
\alias{plotCR}
\title{Plots confidence region estimates}
\usage{
plotCR(estimates)
}
\arguments{
\item{estimates}{See the output of \code{pointIdentifiedCR}.}
}
\description{
A convenience function that creates a plot of each parameter estimate.
}
