% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{pointIdentifiedCR}
\alias{pointIdentifiedCR}
\title{TODO}
\usage{
pointIdentifiedCR(
  ssSize,
  numSubsamples,
  pointEstimate,
  numFreeAttrs,
  groupIDs,
  dataArray,
  optimizeScoreArgs,
  options = list()
)
}
\arguments{
\item{ssSize}{TODO}

\item{numSubsamples}{TODO}

\item{pointEstimate}{TODO}

\item{numFreeAttrs}{TODO}

\item{groupIDs}{TODO}

\item{dataArray}{TODO}

\item{optimizeScoreArgs}{TODO}

\item{options}{TODO}
}
\value{
TODO
}
\description{
TODO
}
