% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{polyeval}
\alias{polyeval}
\title{Evaluate polynomial}
\usage{
polyeval(p, x)
}
\arguments{
\item{p}{The (ordered) coefficients of the polynomial, given as a vector
whose first element is the constant term.}

\item{x}{The point at which the polynomial is evaluated.}
}
\value{
The value of the polynomial
}
\description{
Evaluate the given polynomial at the given point. The function is vectorized
on its second argument.
}
\examples{
polyeval(c(1, -1, 2, -3), 0.5)
}
\keyword{internal}
