% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{sampleBootstrap}
\alias{sampleBootstrap}
\title{Randomly subsample data array}
\usage{
sampleBootstrap(ssSize, groupIDs, dataArray, withReplacement)
}
\arguments{
\item{ssSize}{The size of the desired market subset. Must not be greater than
the original number of markets. Set to \code{NULL} to sample as many
markets as there are in the original set; this is how this function is
supposed to be used in the cube-root method.}

\item{groupIDs}{The output of \code{makeGroupIDs}.}

\item{dataArray}{The output of \code{CdataArray}. Set to \code{NULL} to
generate only the \code{$selectedGroups} and \code{$indices} elements of
the result.}

\item{withReplacement}{A boolean indicating whether to sample with
replacement, as in the cube-root method, or without, as in the
point-identified method.}
}
\value{
A list with members:
\tabular{ll}{
\code{$ssDataArray} \tab An array with the columns of \code{dataArray}
belonging to the selected subset of markets. \cr
\code{$marketIdxs} \tab A vector of length \code{ssSize}, containing the
indices of the selected markets. \cr
\code{$ineqIdxs} \tab A vector of length equal to the number of columns of
the element \code{$ssDataArray}, containing the indices of the selected
inequalities. It has the property that, if \code{r} is the return value
of this function, then \code{dataArray[, r$ineqIdxs]} is equal to
\code{r$ssDataArray}.
}
}
\description{
Generates a reduced data array, with columns corresponding to a randomly
selected subset of the original markets.
}
\keyword{internal}
