% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalities.R
\name{Cineqmembers}
\alias{Cineqmembers}
\title{Compute inequality members for all markets}
\usage{
Cineqmembers(mate)
}
\arguments{
\item{mate}{TODO}
}
\value{
TODO A list of 4-tuples containing indices. For that structure, see
CineqmembersSingle
}
\description{
Computes the upstream and downstream indices for each inequality term for all
markets.
}
