% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{CmatchMatrices}
\alias{CmatchMatrices}
\title{Calculate optimal matchings for all markets}
\usage{
CmatchMatrices(payoffMatrices, quotasU, quotasD)
}
\arguments{
\item{payoffMatrices}{A list of arrays, one for each market. Each array
should have dimension \code{(noD[mIdx], noU[mIdx])}, and its element
indexed by \code{[dIdx, uIdx]} should give the value of the payoff function
for that downstream-upstream pair. See also the return value of the
function \code{evaluatePayoffMatrices}.}

\item{quotasU, quotasD}{Vectors of quotas for the upstreams and downstreams
respectively. See also the arguments of the function
\code{generateAssignmentMatrix}.}
}
\value{
A list of match matrices, one for each market. See the return value
of \code{generateAssignmentMatrix} and \code{importMatched}.
}
\description{
Finds the optimal matching for all markets for the given payoffs and quotas.
}
