% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{generateRandomSubsample}
\alias{generateRandomSubsample}
\title{Randomly subsample data array}
\usage{
generateRandomSubsample(ssSize, groupIDs, dataArray)
}
\arguments{
\item{ssSize}{The size of the market subset. Must not be greater than the
number of markets.}

\item{groupIDs}{The output of \code{makeGroupIDs}.}

\item{dataArray}{The output of \code{CdataArray}.}
}
\value{
An array with the columns of \code{dataArray} belonging to the
selected subset of markets.
}
\description{
Generates a reduced data array, with columns corresponding to a randomly
selected subset of the original markets.
}
\keyword{internal}
