% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{loglikelihoodVarBeta}
\alias{loglikelihoodVarBeta}
\title{ROT log-likelihood with variable beta}
\usage{
loglikelihoodVarBeta(y, x, par)
}
\arguments{
\item{y}{TODO}

\item{x}{TODO}

\item{par}{TODO}
}
\value{
TODO
}
\description{
ROT log-likelihood with variable beta
}
\keyword{internal}
