% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{makeBootstrapObjFun}
\alias{makeBootstrapObjFun}
\title{Create objective function used in Cattaneo's bootstrap}
\usage{
makeBootstrapObjFun(
  fullDataArray,
  sampleDataArray,
  betaEst,
  H,
  coefficient1 = 1,
  objSign = -1
)
}
\arguments{
\item{fullDataArray}{The full data array.}

\item{sampleDataArray}{The sample data array.}

\item{betaEst}{The estimate of \eqn{\beta}. This should have been obtained
by calling \code{optimizeScoreFunction} on \code{fullDataArray}.}

\item{H}{The matrix \eqn{H}, as defined in Cattaneo's paper. Should have
dimensions \eqn{d \times d}, where \eqn{d} is the number of free
parameters.}

\item{coefficient1}{The first coefficient in the extended vector of
parameters. Should typically be \code{1} (the default) or \code{-1}.}

\item{objSign}{Use \code{-1} (the default) when passing the result to an
optimization routine that can only minimize, such as
\code{DEoptim::DEoptim}. Otherwise, use \code{1}.}
}
\value{
A function taking a vector of length \code{noAttr-1} (the parameters)
and returning a real number.
}
\description{
Creates an objective function from the given data, in a form suitable for
passing to an optimization routine. This function is used internally in the
implementation of Cattaneo's bootstrap method.
}
\details{
Let:
\enumerate{
\item \eqn{g_{X}(\beta)} be the score function defined the by data array
\eqn{X} (see \code{makeScoreObjFun}).
\item \eqn{\hat{\beta}} be the estimate of \eqn{\beta}, obtained by
maximizing \eqn{g_{X}}.
\item \eqn{H} be the H-matrix defined in Cattaneo's paper.
\item \eqn{q_{H,\hat{\beta}}(\beta)} be the quadratic form:
\eqn{\frac{1}{2} (\beta - \hat{\beta})^T H (\beta - \hat{\beta})}
}
The function created is:
\eqn{B(\beta) = \mathrm{objSign} (
    g_{X_\mathrm{sample}}(\beta) -
    g_{X_\mathrm{full}}(\beta) -
    q_{H,\hat{\beta}}(\beta)
  )}.

See \code{makeScoreObjFun} for the definition of \eqn{\beta} and
\eqn{\mathrm{objSign}}.
}
