% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R
\name{makeDistanceMatrices}
\alias{makeDistanceMatrices}
\title{TODO}
\usage{
makeDistanceMatrices(attributeMatricesUp, attributeMatricesDn)
}
\arguments{
\item{attributeMatricesUp}{TODO}

\item{attributeMatricesDn}{TODO}
}
\value{
TODO
}
\description{
TODO
}
