% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{makeHplugin}
\alias{makeHplugin}
\title{Create estimator for H matrix}
\usage{
makeHplugin(dataArray, betaEst, h)
}
\arguments{
\item{dataArray}{TODO}

\item{betaEst}{A vector of length \code{d}, whose value is the estimate of
\eqn{\beta}, obtained from maximizing the score function corresponding to
\code{dataArray}.}

\item{h}{TODO}
}
\value{
TODO
}
\description{
Creates the estimator matrix H_n, which is used in Cattaneo's bootstrap
method. Uses the plug-in (kernel) method.
}
\keyword{internal}
