% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{makeScoreObjFun}
\alias{makeScoreObjFun}
\title{Create objective score function}
\usage{
makeScoreObjFun(dataArray, coefficient1 = 1, objSign = -1)
}
\arguments{
\item{dataArray}{See the result of the function \code{CdataArray}.}

\item{coefficient1}{The first coefficient in the extended vector of
parameters. Should typically be \code{1} (the default) or \code{-1}.}

\item{objSign}{Use \code{-1} (the default) when passing the result to an
optimization routine that can only minimize, such as
\code{DEoptim::DEoptim}. Otherwise, use \code{1}.}
}
\value{
A function taking a vector of length \code{noAttr-1} (the parameters)
and returning an scalar (the score).
}
\description{
Creates a score function from the given data array, in a form suitable for
passing to an optimization routine. This function computes the score (number
of inequalities satisfied, normalized on their total number) for a given
parameter vector (\eqn{\beta}).
}
\details{
\loadmathjax

The function created is:
\mjsdeqn{g(\beta) = s \frac{1}{n} \sum_{k=1}^{n} \mathbb{1} \left( a_k^T \tilde{\beta} \right)}
where
\itemize{
\item \mjseqn{c_1} is the first coefficient of the extended vector of
parameters.
\item \mjseqn{s} is \mjseqn{-1} if the function has to be passed to a
minimization method, \mjseqn{1} otherwise.
\item \mjseqn{n} is the number of inequalities/\code{dataArray} rows..
\item \mjseqn{\beta} is the vector of free parameters, of length \code{noAttr-1}.
\item \mjseqn{\tilde{\beta} = (c_1, \beta)} is the extended vector of
parameters.
\item \mjseqn{a_k} is the \mjseqn{k}-th column of \code{dataArray}.
\item \mjseqn{\mathbb{1}(u)} is \mjseqn{1} if \mjseqn{u \geq 0}, \mjseqn{0}
otherwise.
}
}
