% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{mergeOptions}
\alias{mergeOptions}
\title{TODO}
\usage{
mergeOptions(options, defaultOptions)
}
\arguments{
\item{options}{TODO}

\item{defaultOptions}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\keyword{internal}
