% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{pointIdentifiedCR}
\alias{pointIdentifiedCR}
\title{TODO}
\usage{
pointIdentifiedCR(
  dataArray,
  groupIDs,
  pointEstimate,
  ssSize,
  numSubsamples,
  optimizeScoreArgs,
  options = NULL
)
}
\arguments{
\item{dataArray}{TODO}

\item{groupIDs}{TODO}

\item{pointEstimate}{TODO}

\item{ssSize}{TODO}

\item{numSubsamples}{TODO}

\item{optimizeScoreArgs}{TODO}

\item{options}{TODO}
}
\value{
TODO
}
\description{
TODO
}
