% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{rot}
\alias{rot}
\title{ROT calculation of bandwidth}
\usage{
rot(y, x, k, betaEst = NULL, debugLogging = FALSE)
}
\arguments{
\item{y}{TODO}

\item{x}{TODO}

\item{k}{TODO}

\item{betaEst}{TODO}
}
\value{
TODO
}
\description{
ROT calculation of bandwidth
}
\keyword{internal}
